Move checkpoint duty functions to its own module
