+++ Cargo.toml
@@ -138,8 +138,8 @@ num_enum = "0.7"
 parking_lot = "0.12.3"
 paste = "1.0"
 rand = "0.8.5"
-rand_core = "0.6"
 rand_chacha = "0.3.1"
+rand_core = "0.6"
 reqwest = { version = "0.12.4", features = ["json"] }
 reth = { git = "https://github.com/alpenlabs/reth.git", rev = "v1.0.3-alpen.1" }
 reth-basic-payload-builder = { git = "https://github.com/alpenlabs/reth.git", rev = "v1.0.3-alpen.1" }
