+++ crates/consensus-logic/src/duty/mod.rs
@@ -1,4 +1,5 @@
 pub mod block_assembly;
+pub(crate) mod checkpoint;
 pub mod extractor;
 pub mod types;
 pub mod worker;
