+++ provers/sp1/guest-cl-stf/src/vks.rs
@@ -1,1 +1,3 @@
-pub const GUEST_EVM_EE_STF_ELF_ID: &[u32; 8] = &[0, 0, 0, 0, 0, 0, 0, 0];
+pub const GUEST_EVM_EE_STF_ELF_ID: &[u32; 8] = &[
+    1566374620, 1388804633, 1390588015, 1987002781, 1646314853, 378513385, 403362879, 673619552,
+];
