+++ provers/sp1/guest-l1-batch/src/vks.rs
@@ -1,1 +1,3 @@
-pub const GUEST_BTC_BLOCKSPACE_ELF_ID: &[u32; 8] = &[0, 0, 0, 0, 0, 0, 0, 0];
+pub const GUEST_BTC_BLOCKSPACE_ELF_ID: &[u32; 8] = &[
+    185885140, 800202341, 769337952, 994699305, 1179433429, 973199742, 298690500, 1273822492,
+];
